Updated examples to use Slick 3.0
Dropped package structure
