+++ chapter-04/build.sbt
@@ -1,6 +1,6 @@
 name := "essential-slick-chapter-04"
 
-version := "1.0"
+version := "3.0"
 
 scalaVersion := "2.11.6"
 
@@ -19,8 +19,14 @@ scalacOptions ++= Seq(
 
 libraryDependencies ++= Seq(
   "org.scala-lang"      % "scala-reflect"   % scalaVersion.value,
-  "com.typesafe.slick" %% "slick"           % "2.1.0",
+  "com.typesafe.slick" %% "slick"           % "3.0.0",
   "com.h2database"      % "h2"              % "1.4.185",
   "ch.qos.logback"      % "logback-classic" % "1.1.2",
   "joda-time"           % "joda-time"       % "2.6",
   "org.joda"            % "joda-convert"    % "1.2")
+
+initialCommands in console := """
+  |import slick.driver.H2Driver.api._
+  |import StructureExample._
+  |StructureExample.main(Array())
+""".trim.stripMargin
