+++ chapter-04/src/main/resources/application.conf
@@ -1,0 +1,6 @@
+chapter04 = {
+  connectionPool = disabled
+  url = "jdbc:h2:mem:chapter04"
+  driver = "org.h2.Driver"
+  keepAliveConnection = true
+}
